change to double

+++ include/request_generator.hpp
@@ -40,7 +40,7 @@ class RequestGenerator {
 
     Requests generate(uint64_t currentTimeOfDay);
 
-    static uint64_t getTimeMultiplier(uint64_t currentTimeOfDay);
+    static double getTimeMultiplier(uint64_t currentTimeOfDay);
 
    private:
     std::uniform_int_distribution<uint64_t> dropoffDist;
